requirements.txt
streamlit>=1.31.0
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.3.2
pandas>=2.2.0